workflow-orderprocessing-python/services/orderprocessing/requirements.txt
dapr-ext-workflow >= 0.2.0
flask >= 3.0.0
